WICKET-3723 Introduce StorageSettings for the settings related to default inmemory and disk storages
Make the type of maxSizePerSession (the size of the file where session data is written) strongly typed - Bytes.



git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@1128302 13f79535-47bb-0310-9956-ffa450edef68

+++ wicket-core/src/main/java/org/apache/wicket/pageStore/PageWindowManager.java
@@ -73,7 +73,7 @@ public class PageWindowManager implements Serializable
 	 * Maximum page size. After this size is exceeded, the pages will be saved starting at the
 	 * beginning of file.
 	 */
-	private final int maxSize;
+	private final long maxSize;
 
 	/**
 	 * 
@@ -464,7 +464,7 @@ public class PageWindowManager implements Serializable
 	 *            maximum page size. After this size is exceeded, the pages will be saved starting
 	 *            at the beginning of file
 	 */
-	public PageWindowManager(int maxSize)
+	public PageWindowManager(long maxSize)
 	{
 		this.maxSize = maxSize;
 	}
